requirements.txt
fastapi
fastapi-users[beanie]
uvicorn
python-dotenv
beanie